ogb_examples/nodeproppred/unimp/requirements.txt
paddlepaddle_gpu==1.8.3.post107
torch==1.5.1
tqdm==4.31.1
six==1.12.0
numpy==1.19.1
ogb==1.2.1